优化框架结构 增加核心代码

+++ app/src/main/java/com/echo/multidownloader/entities/ThreadInfo.java
@@ -3,9 +3,9 @@ package com.echo.multidownloader.entities;
 public class ThreadInfo {
 	private int id;
 	private String url;
-	private int start;
-	private int end;
-	private int finished;
+	private long start;
+	private long end;
+	private long finished;
 	
 	public ThreadInfo() {
 	}
@@ -17,7 +17,7 @@ public class ThreadInfo {
 	 *@param end
 	 *@param finished
 	 */
-	public ThreadInfo(int id, String url, int start, int end, int finished) {
+	public ThreadInfo(int id, String url, long start, long end, long finished) {
 		this.id = id;
 		this.url = url;
 		this.start = start;
@@ -45,32 +45,26 @@ public class ThreadInfo {
 		this.url = url;
 	}
 
-	public int getStart()
+	public long getStart()
 	{
 		return start;
 	}
 
-	public void setStart(int start)
+	public void setStart(long start)
 	{
 		this.start = start;
 	}
 
-	public int getEnd()
+	public long getEnd()
 	{
 		return end;
 	}
 
-	public void setEnd(int end)
-	{
-		this.end = end;
-	}
+	public void setEnd(long end) {this.end = end;}
 
-	public int getFinished()
-	{
-		return finished;
-	}
+	public long getFinished() {return finished;}
 
-	public void setFinished(int finished)
+	public void setFinished(long finished)
 	{
 		this.finished = finished;
 	}
